initramfs-framework: support mounting rootfs in modules
The "finish" script always expected to mount the rootfs itself if the
"root" parameter was set. Now it checks first whether the rootfs
directory is still empty (using /dev as indicator, because that was
already expected to be present in a supported rootfs) and skips that
step if non-empty.

This allows modules to implement their own rootfs handling.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>