QemuRunner: avoid tainting os.environ
That a utility function permanently changes the process environment is
bad style and leads to subtle, hard to debug problems.

For example, we had one oe-selftest which used runqemu() with an
override for DEPLOY_DIR_IMAGE. Another test then just called runCmd()
and ended up passing the wrong DEPLOY_DIR_IMAGE set earlier in
os.environ.

The approach used here is to extend the launch command such that 'env'
sets the environment variables. The specific values here should be
safe to use this way, even without extra quoting. This approach is
simple and has the advantage that the existing log.info('launchcmd=')
output includes the environment variables.

A more complex approach would be to pass a modified environment hash
to the subprocess module.

[YOCTO #11443]

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>